Refactor CLI corpus commands into submodules
