+++ badges/ripr.json
@@ -1,6 +1,6 @@
 {
   "color": "orange",
   "label": "ripr",
-  "message": "12576",
+  "message": "12600",
   "schemaVersion": 1
 }
